edge case implemented

+++ NamingClient/src/main/java/Client/Threads/Replicate/ReceiveReplicateFileThread.java
@@ -45,13 +45,14 @@ public class ReceiveReplicateFileThread extends Thread {
                 System.out.println("Receive:\n" + hex(buf));
             }
 
-            // TODO log files dink
-            /*if (localFileSet.contains(fileName) || (fileName.startsWith("log_") && localFileSet.contains(fileName.substring(4, fileName.length() - 8)))) {
+            if (localFileSet.contains(fileName) || (fileName.startsWith("log_") && localFileSet.contains(fileName.substring(4, fileName.length() - 8)))) {
                 Thread sendReplicateFileThread = new SendReplicateFileThread(prevNode, dir, fileName);
+                Thread sendLogFileThread = new SendReplicateFileThread(prevNode, dir, "log_" + fileName + ".txt");
                 sendReplicateFileThread.start();
+                sendLogFileThread.start();
                 File f = new File(dir + fileName);
                 f.delete();
-            }*/
+            }
 
             reader.close();
             fileOut.close();
